// src/store/auth/auth.types.ts
import { RNFirebase } from 'react-native-firebase';

export enum AuthActionTypes {
  LOGIN_WITH_GOOGLE_REQUEST = '@auth/LOGIN_WITH_GOOGLE_REQUEST',
  LOGIN_WITH_GOOGLE_SUCCESS = '@auth/LOGIN_WITH_GOOGLE_SUCCESS',
  LOGIN_WITH_GOOGLE_FAILURE = '@auth/LOGIN_WITH_GOOGLE_FAILURE',

  LOGIN_WITH_FACEBOOK_REQUEST = '@auth/LOGIN_WITH_FACEBOOK_REQUEST',
  LOGIN_WITH_FACEBOOK_SUCCESS = '@auth/LOGIN_WITH_FACEBOOK_SUCCESS',
  LOGIN_WITH_FACEBOOK_FAILURE = '@auth/LOGIN_WITH_FACEBOOK_FAILURE',

  LOGIN_WITH_EMAIL_AND_PASSWORD_REQUEST = '@auth/LOGIN_WITH_EMAIL_AND_PASSWORD_REQUEST',
  LOGIN_WITH_EMAIL_AND_PASSWORD_SUCCESS = '@auth/LOGIN_WITH_EMAIL_AND_PASSWORD_SUCCESS',
  LOGIN_WITH_EMAIL_AND_PASSWORD_FAILURE = '@auth/LOGIN_WITH_EMAIL_AND_PASSWORD_FAILURE',

  SIGN_UP_WITH_EMAIL_AND_PASSWORD_REQUEST = '@auth/SIGN_UP_WITH_EMAIL_AND_PASSWORD_REQUEST',
  SIGN_UP_WITH_EMAIL_AND_PASSWORD_SUCCESS = '@auth/SIGN_UP_WITH_EMAIL_AND_PASSWORD_SUCCESS',
  SIGN_UP_WITH_EMAIL_AND_PASSWORD_FAILURE = '@auth/SIGN_UP_WITH_EMAIL_AND_PASSWORD_FAILURE',

  SIGN_OUT_REQUEST = '@auth/SIGN_OUT_REQUEST',
  SIGN_OUT_SUCCESS = '@auth/SIGN_OUT_SUCCESS',
  SIGN_OUT_FAILURE = '@auth/SIGN_OUT_FAILURE',

  SEND_PASSWORD_RESET_EMAIL_REQUEST = '@auth/SEND_PASSWORD_RESET_EMAIL_REQUEST',
  SEND_PASSWORD_RESET_EMAIL_SUCCESS = '@auth/SEND_PASSWORD_RESET_EMAIL_SUCCESS',
  SEND_PASSWORD_RESET_EMAIL_FAILURE = '@auth/SEND_PASSWORD_RESET_EMAIL_FAILURE',

  CLEAR_ERRORS = '@auth/CLEAR_ERRORS',
}

export interface AuthState {
  readonly loading: boolean;
  readonly authenticated: boolean;
  readonly errors: Error[];
  readonly user: RNFirebase.auth.OrNull<RNFirebase.User>;
}